Don't declare locals too soon.
Without this, the following fails:

    $ cat scratch.fnl
    (fn [abcd] (let [abcd abcd] abcd))
    $ make && fennel --plugin src/linter.fnl --compile scratch.fnl

This is because both references to abcd get counted as a reference to
the let-bound abcd, and none get counted to the function
parameter. With this patch applied, the third reference applies to the
function parameter, and the fourth applies to the let-bound local.

This fix works, but it's less than ideal. It turns a bunch of local
names into thunks to avoid declaring them too early. There is a mutual
dependency between the left-hand side and the compilation of the right
hand side due to the compilation requiring an opts.target value which
tells the compiler what to bind the results to. The thunks work around
the problem, but in an ugly way.